// shared/types/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      comments: {
        Row: {
          author: string | null
          content: string
          created_at: string
          id: number
          post_id: string
        }
        Insert: {
          author?: string | null
          content: string
          created_at?: string
          id?: number
          post_id?: string
        }
        Update: {
          author?: string | null
          content?: string
          created_at?: string
          id?: number
          post_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "comments_post_id_fkey"
            columns: ["post_id"]
            isOneToOne: false
            referencedRelation: "post_summary"
            referencedColumns: ["post_id"]
          },
          {
            foreignKeyName: "comments_post_id_fkey"
            columns: ["post_id"]
            isOneToOne: false
            referencedRelation: "posts"
            referencedColumns: ["post_id"]
          },
        ]
      }
// ...
    }
// ...
    Functions: {
      get_all_active_posts: {
        Args: Record<PropertyKey, never>
        Returns: {
          content: string
          created_at: string
          id: number
          is_deleted: boolean
          title: string
        }[]
      }
      get_channels: {
        Args: Record<PropertyKey, never>
        Returns: {
          channel: string
        }[]
      }
      get_comments: {
        Args: { p_page?: number; p_per_page?: number; p_post_id: string }
        Returns: {
          author_avatar: string
          author_name: string
          content: string
          created_at: string
          id: number
          post_id: string
        }[]
      }
      get_post: {
        Args: { page_number: number; per_page: number }
        Returns: {
          attachments: Json
          avatar: string
          comment_count: number
          content: string
          created_at: string
          is_commented: boolean
          is_liked: boolean
          like_count: number
          name: string
          post_id: string
          title: string
          uid: string
        }[]
      }
      get_post_by_id: {
        Args: { p_post_id: string }
        Returns: {
          attachments: Json
          avatar: string
          comment_count: number
          content: string
          created_at: string
          is_commented: boolean
          is_liked: boolean
          like_count: number
          name: string
          post_id: string
          title: string
          uid: string
        }[]
      }
      get_target_app_version_list: {
        Args: {
          app_platform: Database["public"]["Enums"]["platforms"]
          min_bundle_id: string
        }
        Returns: {
          target_app_version: string
        }[]
      }
      get_update_info_by_app_version: {
        Args: {
          app_platform: Database["public"]["Enums"]["platforms"]
          app_version: string
          bundle_id: string
          min_bundle_id: string
          target_app_version_list: string[]
          target_channel: string
        }
        Returns: {
          id: string
          message: string
          should_force_update: boolean
          status: string
          storage_uri: string
        }[]
      }
      get_update_info_by_fingerprint_hash: {
        Args: {
          app_platform: Database["public"]["Enums"]["platforms"]
          bundle_id: string
          min_bundle_id: string
          target_channel: string
          target_fingerprint_hash: string
        }
        Returns: {
          id: string
          message: string
          should_force_update: boolean
          status: string
          storage_uri: string
        }[]
      }
      increase_user_love_hero_status: {
        Args: { p_hero_id: number; p_uid: string } | { uid: string }
        Returns: undefined
      }
      notify: {
        Args: {
          p_body: string
          p_image_url?: string
          p_screen?: string
          p_title: string
          p_user_id: string
        }
        Returns: Json
      }
      send_firebase_message: {
        Args: {
          p_body: string
          p_image_url?: string
          p_screen?: string
          p_title: string
          p_user_id: string
        }
        Returns: undefined
      }
      toggle_like: {
        Args: { p_post_id: string } | { p_post_id: string; p_user_id: string }
        Returns: Json
      }
    }
    Enums: {
      attachment_type: "image" | "video" | "file" | "youtube_url" | "audio"
      hero_roles_enum: "ad" | "ap" | "tank" | "sp"
      platforms: "ios" | "android"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never